move label component to parent folder
