+++ components/index.ts
@@ -4,6 +4,7 @@ export * from './button/Button';
 export { default as Button } from './button/Button';
 export * from './icon/Icon';
 export { default as Icon } from './icon/Icon';
+export { Label } from './label/Label';
 export { Heading } from './typography/Heading';
-export { Label } from './typography/Label';
 export { Paragraph } from './typography/Paragraph';
+
