+++ components/input-group/InputFieldGroup.tsx
@@ -3,22 +3,22 @@ import React, { useId } from 'react';
 import Icon from '../icon/Icon';
 import { EIConTypes } from '../icon/icon.enum';
 import { TIconType } from '../icon/icon.interface';
-import { Label } from '../typography/Label';
+import { Label } from '../label/Label';
 
 export interface IInputFieldGroupProps {
   id: string;
   text?: string;
   placeholder: string;
   as:
-    | 'text'
-    | 'password'
-    | 'email'
-    | 'number'
-    | 'tel'
-    | 'url'
-    | 'search'
-    | 'file'
-    | 'hidden';
+  | 'text'
+  | 'password'
+  | 'email'
+  | 'number'
+  | 'tel'
+  | 'url'
+  | 'search'
+  | 'file'
+  | 'hidden';
   onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
   state: 'default' | 'error' | 'success';
   required?: boolean;
@@ -60,7 +60,7 @@ export const InputFieldGroup = ({
             'focus:outline focus:outline-2',
             state === 'default' && 'focus:border-none focus:outline-violet-600',
             state === 'error' &&
-              'outline outline-2 outline-red hover:border-red',
+            'outline outline-2 outline-red hover:border-red',
             state == 'success' && '',
           )}
         />
